Reorganizing into separate files for each processor. Removed unused DocsplitChaining processor for now.
