+++ lib/docsplit-paperclip-processor.rb
@@ -17,105 +17,8 @@ module Paperclip
       File.expand_path(@src.path)
     end
   end
-
-  class DocsplitChaining < Processor
-    attr_accessor :options, :attachment
-
-    def initialize(file, options = {}, attachment = nil)
-      super
-      @options    = options
-      @attachment = attachment
-    end
-
-    def make
-      attachment.to_file(options[:from_style] || :original)
-    end
-  end
-
-  class DocsplitPdf < DocsplitProcessor
-    def make
-      begin
-        dst_dir = Dir.tmpdir
-        dst_path = File.join(dst_dir, "#{@basename}.pdf")
-
-        if pdf_format?
-          dst_path = File.join(dst_dir, "_#{@basename}.pdf")
-          FileUtils.copy_file(src_path, dst_path)
-        else
-          Docsplit.extract_pdf(src_path, :output => dst_dir)
-        end
-      rescue Exception => e
-        raise Paperclip::Error, "There was an error converting #{@basename} to pdf"
-      end
-      File.open(dst_path)
-    end
-
-    def pdf_format?
-      file_magic = FileMagic.new
-      type = file_magic.file(src_path)
-      file_magic.close
-      type =~ /pdf/i
-    end
-  end
-
-  class DocsplitImage < DocsplitProcessor
-    def make
-      begin
-        @dst_path = Dir.tmpdir
-        @pages    = @options[:pages] || [1]
-        @options  = @options.merge(:output => @dst_path)
-
-        Docsplit.extract_images(src_path, @options)
-      rescue Exception => e
-        raise Paperclip::Error, "There was an error extracting images from #{@basename}"
-      end
-      
-      destination_file
-    end
-
-    def destination_file
-      File.open(File.join(@dst_path, "#{@basename}_#{@pages.first}.#{@options[:format]}"))
-    end
-  end
-
-  class DocsplitText < DocsplitProcessor
-    def make
-      begin
-        @dst_path = Dir.tmpdir
-        @pages    = @options[:pages] || [1]
-        @options  = @options.merge(:output => @dst_path)
-
-        Docsplit.extract_text(src_path, @options)
-      rescue Exception => e
-        raise Paperclip::Error, "There was an error extracting text from #{@basename}"
-      end
-
-      if @options[:full_text_column]
-        # Bypassing callbacks to save full text. See Paperclip issue #671:
-        # https://github.com/thoughtbot/paperclip/issues/671
-        ar_model = @attachment.instance
-        ar_model[@options[:full_text_column]] = full_text
-        ar_model.run_callbacks(:save) { false }
-
-        # This would be the preferred method of saving this text.
-        # @attachment.instance.update_attribute(@options[:full_text_column], full_text)
-      end
-
-      destination_file
-    end
-
-    def destination_file
-      File.open(File.join(@dst_path, "#{@basename}.txt"))
-    end
-
-    def full_text
-      full_text = String.new
-
-      destination_file.each do |line|
-        full_text += line
-      end
-
-      full_text
-    end
-  end
 end
+
+require 'processors/docsplit_image'
+require 'processors/docsplit_pdf'
+require 'processors/docsplit_text'
